Revert "chore: refactor didroom java class (#333)"
This reverts commit eb16abcb3d115ac3bd987fcb2fd6dde6be2447c9.
