+++ android/app/capacitor.build.gradle
@@ -20,7 +20,6 @@ dependencies {
     implementation project(':capacitor-keyboard')
     implementation project(':capacitor-preferences')
     implementation project(':capacitor-status-bar')
-    implementation project(':capacitor-native-settings')
 
 }
