+++ android/local.properties.example
@@ -1,0 +1,1 @@
+sdk.dir=/Users/<HOME_NAME>/Library/Android/sdk
