+++ fastlane/AppFile
@@ -1,2 +1,3 @@
-package_name("com.forkbomb.wallet") 
 json_key_file "play-store-credentials.json"
+package_name "eu.forkbomb.didroom"
+
